Update README.md
Hello,
This Project is done by 
D.Sai Chethan,
B.Sai Bhanu,
P.Surya Srikar.

**About Project**
This Project is called Kisaan-Konnect, as the name says. This is a e-platform for farmers to sell their produce. This way farmers have direct connection with retailers without any intermediate person. This app is very similar to OLX, to make it unique we want to add certain changes which we haven't done yet but we want to add a machine learning model to this app where it shows the current market price of a certain product and also predict it's market price in future. Also, similar to OLX, retailer could bargain for suitable price to which the farmer can agree and the mode of communication between farmer and retailer for now is just contact no., which will be displayed on product page.

**What We have Done**
We used flutter in developing the app and cloud firestore for storing the user and product data. There are two models, a User model and a Item model. Each User has collection of items he has sold till now. Each Item has details of it's owner linked to it. Our app starts with authentication page, One can register with email and password. Here are few features of our app
1.User Authentication(register, login, keep the user signed in, sign out).
2.Home Page with list of items added and its features
3.Ability to add new items with name, price, location and date of harvest.
4.My ads section for items added by the logged-in user.
5.Implemented Search Bar
6.User details like name, contact number, location saved and retrieved to firestore database with option to update.

**What else more we want to add**
1.As mentioned we want to create a item page with description about the item, image and also it's owners details and a option to bargain
2.A Machine Learning Model which can predict the market price of a product in future